Boxer panel: set timings for 1024x600@54Hz
In order for the DSS scaler to be able to downscale 720p video to our
panel resolution, the DSS FCLK needs to be at least 4x the pixel clock.
Since the maximum FCLK supported by the OMAP3 DSS is 173 MHz, we need to
reduce our pixel clock to 1/4 of this value or less for 720p downscaling
to work.  43.2 MHz is the maximum supported pixel clock which meets this
criterion.

At 43.2 MHz, we can theoretically carry up to 70 1024x600 frames per
second.  As a practical matter, though, the panel needs synchronization,
which eats into the available bandwidth for frame data.  Because the PVR
framebuffer HAL configuration (powervr.ini) only supports specifying
integer values for the refresh rate, we need to choose timings which
allow an integer rate.

With those constraints in mind, the only refresh rates we'd want to use
are 54 Hz, 60 Hz, and 64 Hz.  We'd prefer a higher refresh rate, but
running the panel at 60 Hz causes a distracting strobing effect (related
to the backlight flicker frequency?), and despite the math, I haven't
found any timings which the panel will accept for 64 Hz.  Therefore, we
set panel timings for 1024x600@54Hz.  (We're in decent company here;
tuna sets a refresh rate around 58 Hz, and crespo around 56 Hz.)